feat: module regions #13
