+++ flutter_dex/lib/main.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_dex/components/homeButtom.dart';
-import 'package:flutter_dex/components/bottomNavBar.dart';
 import 'package:flutter_dex/homepage/homePage.dart';
 import 'package:flutter_dex/quiz/telaQuiz.dart';
+import 'package:flutter_dex/regions/telaRegion.dart';
 
 void main() {
   runApp(MyApp());
@@ -23,27 +22,6 @@ class MyApp extends StatelessWidget {
         'telaregions': (context) => TelaRegion(),
         'telaquiz': (context) => TelaQuiz(),
       },
-    );
-  }
-}
-
-
-class TelaRegion extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Dex basics'),
-      ),
-      body: const Center(
-        child: Text(
-          'Página Regioes',
-          style: TextStyle(fontSize: 24),
-        ),
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: DexButtomHome(),
-      bottomNavigationBar: DexAppBar(),
     );
   }
 }
